Move is_fi_a2a_supported to utils.common
